+++ samples/jt-808-server-sample-bare-boot3/src/main/java/io/github/hylexus/jt/jt808/samples/boot3/bare/Jt808ServerSampleBareBoot3Application.java
@@ -1,0 +1,17 @@
+package io.github.hylexus.jt.jt808.samples.boot3.bare;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * @author hylexus
+ */
+@SpringBootApplication
+public class Jt808ServerSampleBareBoot3Application {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Jt808ServerSampleBareBoot3Application.class, args);
+    }
+
+}
+